# packages/contracts/openapi/ai-internal.yaml
openapi: 3.1.0
info:
  title: TravelMate Internal AI API
  version: 1.1.0
servers:
  - url: http://127.0.0.1:8000/internal/v1/ai
paths:
  /chat:
    post:
      operationId: chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Grounded travel reply
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
  /itineraries/generate:
    post:
      operationId: generateItinerary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItineraryRequest'
      responses:
        '200':
          description: Generated plan or clarification questions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItineraryResponse'
  /suggest-places:
    post:
      operationId: suggestPlaces
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuggestPlacesRequest'
      responses:
        '200':
          description: Catalog-backed and geocoded place suggestions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceSuggestionResponse'
components:
  schemas:
    ChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 1000
        history:
          type: array
          maxItems: 20
          items:
            $ref: '#/components/schemas/ChatHistoryMessage'
        tripContext:
          $ref: '#/components/schemas/TripContext'
    ChatHistoryMessage:
      type: object
      required: [role, content]
      properties:
        role:
          enum: [user, assistant]
        content:
          type: string
          minLength: 1
          maxLength: 2000
    TripContext:
      type: object
      properties:
        destination:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        budgetVnd:
          type: integer
          minimum: 0
        numPeople:
          type: integer
          minimum: 1
          maximum: 50
    ChatResponse:
      type: object
      required: [reply, isOutOfScope, suggestedQuestions, provider, resetContext]
      properties:
        reply:
          type: string
        isOutOfScope:
          type: boolean
        suggestedQuestions:
          type: array
          maxItems: 3
          items:
            type: string
        provider:
          enum: [mock, gemini, groq, local]
        modelVersion:
          type: [string, 'null']
        resetContext:
          type: boolean
    ItineraryRequest:
      type: object
      properties:
        destination:
          type: [string, 'null']
        startDate:
          type: [string, 'null']
          format: date
        endDate:
          type: [string, 'null']
          format: date
        durationDays:
          type: [integer, 'null']
          minimum: 1
          maximum: 14
        numPeople:
          type: [integer, 'null']
          minimum: 1
          maximum: 50
        budgetVnd:
          type: [integer, 'null']
          minimum: 0
        preferences:
          type: array
          maxItems: 8
          items:
            type: string
        notes:
          type: [string, 'null']
          maxLength: 1000
    ItineraryResponse:
      type: object
      required: [status, missingFields, questions]
      properties:
        status:
          enum: [needs_clarification, ready]
        missingFields:
          type: array
          items:
            enum: [destination, durationDays, numPeople, budgetVnd]
        questions:
          type: array
          maxItems: 4
          items:
            type: string
        plan:
          oneOf:
            - $ref: '#/components/schemas/ItineraryPlan'
            - type: 'null'
        provider:
          type: [string, 'null']
          enum: [mock, gemini, groq, local, null]
    ItineraryPlan:
      type: object
      required: [destination, durationDays, numPeople, summary, days, budget]
      properties:
        destination:
          type: string
        durationDays:
          type: integer
        numPeople:
          type: integer
        summary:
          type: string
        assumptions:
          type: array
          items:
            type: string
        days:
          type: array
          items:
            $ref: '#/components/schemas/ItineraryDay'
        budget:
          type: object
          additionalProperties:
            type: integer
            minimum: 0
    ItineraryDay:
      type: object
      required: [day, title, activities]
      properties:
        day:
          type: integer
        title:
          type: string
        activities:
          type: array
          items:
            $ref: '#/components/schemas/ItineraryActivity'
    ItineraryActivity:
      type: object
      required: [period, kind, title]
      properties:
        period:
          enum: [morning, afternoon, evening]
        kind:
          enum: [visit, meal, rest, travel, free_time]
        title:
          type: string
        placeId:
          type: [string, 'null']
        placeName:
          type: [string, 'null']
        notes:
          type: [string, 'null']
    SuggestPlacesRequest:
      type: object
      required: [city]
      properties:
        city:
          type: string
          minLength: 1
          maxLength: 120
        type:
          type: [string, 'null']
          maxLength: 60
        specialNote:
          type: [string, 'null']
          maxLength: 500
        count:
          type: integer
          default: 5
          minimum: 1
          maximum: 6
    PlaceSuggestionResponse:
      type: object
      required: [city, suggestions, provider]
      properties:
        city:
          type: string
        suggestions:
          type: array
          items:
            $ref: '#/components/schemas/PlaceSuggestion'
        message:
          type: [string, 'null']
        provider:
          const: catalog
    PlaceSuggestion:
      type: object
      required: [id, name, category, description, reason, address, latitude, longitude, mapUrl, source]
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        description:
          type: string
        reason:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        mapUrl:
          type: string
        source:
          type: string